Theme is correctly loaded and assigned to monaco
